+++ src/core/hooks/use-force-update.hook.ts
@@ -1,0 +1,12 @@
+import * as React from 'react'
+import * as R from 'ramda'
+
+export const useForceUpdate = (): () => void => {
+  const [ _, setState ] = React.useState(0)
+
+  const forceUpdate = () => {
+    setState(R.inc)
+  }
+
+  return forceUpdate
+}